refactor: refactor components folder structure
