--- src/components/Avatars/SummerbudAvatars/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./SummerbudAvatars";
-export * from "./SummerbudAvatars";
