--- src/components/Avatars/index.ts
@@ -1,4 +1,0 @@
-import SummerbudAvatars from "./SummerbudAvatars";
-import type { SummberbudAvatarProps } from "./SummerbudAvatars";
-
-export { SummerbudAvatars, SummberbudAvatarProps };
