--- src/components/Buttons/ThemeToggle/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ThemeToggle";
