--- src/components/Buttons/index.ts
@@ -1,3 +1,0 @@
-import ThemeToggle from "./ThemeToggle";
-
-export { ThemeToggle };
