--- src/components/Footer/index.ts
@@ -1,2 +1,0 @@
-import Footer from "./Footer";
-export { Footer };
