--- src/components/Giscus/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Giscus";
