--- src/components/Image/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Image";
-export * from "./Image";
