--- src/components/PostList/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./PostList";
-export * from "./PostList";
