+++ src/layouts/SiteLayout.astro
@@ -1,8 +1,7 @@
 ---
 import HeadSeo from "./HeadSeo.astro";
-import NavBar from "../components/NavBar";
 import { Props as HeadSeoProps } from "./HeadSeo.astro";
-import { Footer } from "../components/Footer";
+import { Footer, NavBar } from "@/components/index";
 import cn from "clsx";
 
 export interface Props extends HeadSeoProps {
@@ -17,7 +16,11 @@ const props = Astro.props as Props;
     <meta content="width=device-width, initial-scale=1" name="viewport" />
     <meta charset="utf-8" />
     <link href="/prism-atom-dark.css" rel="stylesheet" />
-    <HeadSeo meta={props.meta} canonicalURL={props.canonicalURL} text={props.text} />
+    <HeadSeo
+      meta={props.meta}
+      canonicalURL={props.canonicalURL}
+      text={props.text}
+    />
     <script is:inline>
       // credit: https://github.com/withastro/docs/blob/main/src/layouts/MainLayout.astro
       const html = document.querySelector("html");
