+++ src/pages/thoughts/index.astro
@@ -1,5 +1,5 @@
 ---
-import PostList from "../../components/PostList";
+import { PostList } from "@/components/index";
 import { ArticleFrontMatter } from "../../types/general";
 import SiteLayout, {
   Props as SiteLayoutProps,
